feat: Move library metrics from domain
