+++ src/features/metrics/handlers.go
@@ -4,7 +4,6 @@ import (
 	"log/slog"
 	"strings"
 
-	"github.com/contre95/soulsolid/src/music"
 	"github.com/gofiber/fiber/v2"
 )
 
@@ -128,7 +127,7 @@ func (h *Handler) GetMetadataChartHTML(c *fiber.Ctx) error {
 	lyricsStats, err := h.service.metrics.GetLyricsStats(c.Context())
 	if err != nil {
 		slog.Error("Error getting lyrics stats", "error", err)
-		lyricsStats = music.LyricsStats{}
+		lyricsStats = LyricsStats{}
 	}
 
 	// Calculate percentages
